requirements.txt
Flask==1.1.2
Flask-Migrate==2.5.3
Flask-SQLAlchemy==2.4.3
wheel==0.34.2
psycopg2==2.8.5
pypugjs==5.9.4
